+++ ICD.Connect.Protocol.Network/Ports/Web/AbstractWebPort.cs
@@ -62,10 +62,9 @@ namespace ICD.Connect.Protocol.Network.Ports.Web
 		/// Sends a GET request to the server.
 		/// </summary>
 		/// <param name="relativeOrAbsoluteUri"></param>
-		/// <param name="response"></param>
-		public bool Get(string relativeOrAbsoluteUri, out string response)
+		public WebPortResponse Get(string relativeOrAbsoluteUri)
 		{
-			return Get(relativeOrAbsoluteUri, new Dictionary<string, List<string>>(), out response);
+			return Get(relativeOrAbsoluteUri, new Dictionary<string, List<string>>());
 		}
 
 		/// <summary>
@@ -73,19 +72,17 @@ namespace ICD.Connect.Protocol.Network.Ports.Web
 		/// </summary>
 		/// <param name="relativeOrAbsoluteUri"></param>
 		/// <param name="headers"></param>
-		/// <param name="response"></param>
-		public abstract bool Get(string relativeOrAbsoluteUri, IDictionary<string, List<string>> headers, out string response);
+		public abstract WebPortResponse Get(string relativeOrAbsoluteUri, IDictionary<string, List<string>> headers);
 
 		/// <summary>
 		/// Sends a POST request to the server.
 		/// </summary>
 		/// <param name="relativeOrAbsoluteUri"></param>
 		/// <param name="data"></param>
-		/// <param name="response"></param>
 		/// <returns></returns>
-		public bool Post(string relativeOrAbsoluteUri, byte[] data, out string response)
+		public WebPortResponse Post(string relativeOrAbsoluteUri, byte[] data)
 		{
-			return Post(relativeOrAbsoluteUri, new Dictionary<string, List<string>>(), data, out response);
+			return Post(relativeOrAbsoluteUri, new Dictionary<string, List<string>>(), data);
 		}
 
 		/// <summary>
@@ -94,18 +91,16 @@ namespace ICD.Connect.Protocol.Network.Ports.Web
 		/// <param name="relativeOrAbsoluteUri"></param>
 		/// <param name="headers"></param>
 		/// <param name="data"></param>
-		/// <param name="response"></param>
 		/// <returns></returns>
-		public abstract bool Post(string relativeOrAbsoluteUri, Dictionary<string, List<string>> headers, byte[] data, out string response);
+		public abstract WebPortResponse Post(string relativeOrAbsoluteUri, Dictionary<string, List<string>> headers, byte[] data);
 
 		/// <summary>
 		/// Sends a SOAP request to the server.
 		/// </summary>
 		/// <param name="action"></param>
 		/// <param name="content"></param>
-		/// <param name="response"></param>
 		/// <returns></returns>
-		public abstract bool DispatchSoap(string action, string content, out string response);
+		public abstract WebPortResponse DispatchSoap(string action, string content);
 
 		/// <summary>
 		/// Applies the given device configuration properties to the port.
@@ -348,9 +343,8 @@ namespace ICD.Connect.Protocol.Network.Ports.Web
 		/// <param name="path"></param>
 		private string ConsoleGet(string path)
 		{
-			string output;
-			Get(path, out output);
-			return output;
+			WebPortResponse response = Get(path);
+			return response.DataAsString;
 		}
 
 		/// <summary>
@@ -359,9 +353,8 @@ namespace ICD.Connect.Protocol.Network.Ports.Web
 		/// <param name="path"></param>
 		private string ConsolePost(string path)
 		{
-			string output;
-			Post(path, new byte[0], out output);
-			return output;
+			WebPortResponse response = Post(path, new byte[0]);
+			return response.DataAsString;
 		}
 
 		#endregion